+++ packages/cli/lib/runner.dart
@@ -7,6 +7,7 @@ import 'package:superdeck_builder/superdeck_builder.dart';
 import 'package:superdeck_core/superdeck_core.dart';
 
 import 'src/commands/build_command.dart';
+import 'src/commands/run_command.dart';
 import 'src/commands/setup_command.dart';
 import 'src/utils/constants.dart';
 import 'src/utils/logger.dart';
@@ -23,6 +24,7 @@ class SuperDeckRunner extends CommandRunner<int> {
     Logger? loggerOverride,
     List<DeckBuildPlugin> plugins = const [],
     BuildCommand? buildCommand,
+    RunCommand? runCommand,
     SetupCommand? setupCommand,
   }) : super(cliName, cliDescription) {
     _logger = loggerOverride ?? logger;
@@ -44,6 +46,9 @@ class SuperDeckRunner extends CommandRunner<int> {
 
     addCommand(
       buildCommand ?? BuildCommand(loggerOverride: _logger, plugins: plugins),
+    );
+    addCommand(
+      runCommand ?? RunCommand(loggerOverride: _logger, plugins: plugins),
     );
     addCommand(setupCommand ?? SetupCommand(loggerOverride: _logger));
   }